+++ src/App.js
@@ -3,23 +3,17 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import './styles/App.css';
 
-import Home from './components/Home';
-import Main from './components/Main';
-import SignIn from './pages/SignIn';
-import SignUp from './pages/SignUp';
-import Profile from './pages/Profile';
-import Qna from './pages/Qna';
-import { QnaWrite } from './components/Qna/QnaWrite';
-import { QnaView } from './components/Qna/QnaView';
-
 import Home from './pages/Home';
 import Main from './pages/Main';
 import Notice from './pages/Notice';
+import Qna from './pages/Qna';
 import SignIn from './components/Login/SignIn';
 import SignUp from './components/Login/SignUp';
 import NoticeView from './components/Notice/NoticeView'
 import NoticeWrite from './components/Notice/NoticeWrite';
 import NoticeModify from './components/Notice/NoticeModify';
+import { QnaWrite } from './components/Qna/QnaWrite';
+import { QnaView } from './components/Qna/QnaView';
 
 
 
@@ -32,8 +26,6 @@ function App() {
           <Route path='/signin' element={<SignIn />} />
           <Route path='/signup' element={<SignUp />} />
           <Route path='/home' element={<Home />} />
-          
-          <Route path='/profile' element={<Profile />} />
           
           <Route path='/notice' element={<Notice />} />
           <Route path='/notice/write' element={<NoticeWrite />} />
