+++ src/components/Base/Nav.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../styles/Main.css';
+import '../../styles/Main.css';
 import { Link } from 'react-router-dom';
 
 export default function Nav() {
