+++ src/components/Qna/QnaWrite.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import Header from '../Header'
+import Header from '../Base/Header'
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
